+++ components/ui/3d-card.tsx
@@ -42,10 +42,7 @@ export const CardContainer = ({
   return (
     <MouseEnterContext.Provider value={[isMouseEntered, setIsMouseEntered]}>
       <div
-        className={cn(
-          'py-20 flex items-center justify-center',
-          containerClassName
-        )}
+        className={cn('flex items-center justify-center', containerClassName)}
         style={{
           perspective: '1000px',
         }}
@@ -56,7 +53,7 @@ export const CardContainer = ({
           onMouseMove={handleMouseMove}
           onMouseLeave={handleMouseLeave}
           className={cn(
-            'flex items-center justify-center relative transition-all duration-200 ease-linear',
+            'flex items-stretch justify-between relative transition-all duration-200 ease-linear',
             className
           )}
           style={{
@@ -80,7 +77,7 @@ export const CardBody = ({
   return (
     <div
       className={cn(
-        'h-96 w-96 [transform-style:preserve-3d]  [&>*]:[transform-style:preserve-3d]',
+        '[transform-style:preserve-3d]  [&>*]:[transform-style:preserve-3d]',
         className
       )}
     >